refactor: Refactor sidebar. Add Metadachi svg
