+++ components/sidebar/SidebarItem.tsx
@@ -5,8 +5,8 @@ import { Tooltip } from "@mui/joy"
 import ListItemDecorator from "@mui/joy/ListItemDecorator"
 import ListItemContent from "@mui/joy/ListItemContent"
 import Typography from "@mui/joy/Typography"
-import { routeDictionary } from "@/components/ui/Sidebar"
 import { useRouter } from "next/navigation"
+import { routeDictionary } from "@/components/sidebar/RouteDictionary"
 
 export function SidebarItem({
   selected = false,
